Stop duplicating the StellarId and Customer objects, and move them to a shared folder.
