+++ core/src/test/kotlin/org/stellar/anchor/sep31/Sep31ServiceTest.kt
@@ -27,6 +27,8 @@ import org.stellar.anchor.api.sep.sep31.*
 import org.stellar.anchor.api.sep.sep31.Sep31PostTransactionRequest.Sep31TxnFields
 import org.stellar.anchor.api.sep.sep38.RateFee
 import org.stellar.anchor.api.shared.Amount
+import org.stellar.anchor.api.shared.Customers
+import org.stellar.anchor.api.shared.StellarId
 import org.stellar.anchor.asset.AssetService
 import org.stellar.anchor.asset.ResourceJsonAssetService
 import org.stellar.anchor.auth.JwtService
@@ -35,8 +37,6 @@ import org.stellar.anchor.config.Sep31Config
 import org.stellar.anchor.config.Sep31Config.PaymentType.STRICT_RECEIVE
 import org.stellar.anchor.config.Sep31Config.PaymentType.STRICT_SEND
 import org.stellar.anchor.event.EventPublishService
-import org.stellar.anchor.event.models.Customers
-import org.stellar.anchor.event.models.StellarId
 import org.stellar.anchor.event.models.TransactionEvent
 import org.stellar.anchor.sep31.Sep31Service.*
 import org.stellar.anchor.sep38.PojoSep38Quote
