+++ core/src/test/kotlin/org/stellar/anchor/sep38/Sep38ServiceTest.kt
@@ -20,12 +20,12 @@ import org.stellar.anchor.api.exception.ServerErrorException
 import org.stellar.anchor.api.sep.AssetInfo
 import org.stellar.anchor.api.sep.sep38.*
 import org.stellar.anchor.api.sep.sep38.Sep38Context.*
+import org.stellar.anchor.api.shared.StellarId
 import org.stellar.anchor.asset.ResourceJsonAssetService
 import org.stellar.anchor.config.AppConfig
 import org.stellar.anchor.config.Sep38Config
 import org.stellar.anchor.event.EventPublishService
 import org.stellar.anchor.event.models.QuoteEvent
-import org.stellar.anchor.event.models.StellarId
 
 class Sep38ServiceTest {
   internal class PropertySep38Config : Sep38Config {
