+++ platform/src/main/java/org/stellar/anchor/platform/data/JdbcSep31Transaction.java
@@ -10,7 +10,7 @@ import javax.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 import org.stellar.anchor.api.sep.AssetInfo;
-import org.stellar.anchor.event.models.StellarId;
+import org.stellar.anchor.api.shared.StellarId;
 import org.stellar.anchor.reference.model.StellarIdConverter;
 import org.stellar.anchor.sep31.Sep31Transaction;
 import org.stellar.anchor.util.GsonUtils;
